Move credential fixture to conftest
